search functionily and many more updates

+++ src/components/dashboard/UsersData/UsersData.jsx
@@ -1,20 +1,45 @@
 import { useContext } from "react";
 import { DataContext } from "../../../provider/AppContext";
 import { RiDeleteBin6Fill } from "react-icons/ri";
+import Swal from 'sweetalert2/dist/sweetalert2.js'
+import 'sweetalert2/dist/sweetalert2.css'
+import { AuthContext } from "../../../provider/AuthProvider";
+import { toast } from "react-toastify";
 
 const UsersData = ({ userData, index }) => {
     const { profileImage, name, email, role } = userData;
     const { updateUserRole, deleteUser } = useContext(DataContext)
+    const { user } = useContext(AuthContext)
 
-    const handleUserRoleChange = (e) => {
+    const handleUserRoleChange = async (e) => {
         const role = e.target.value;
-        updateUserRole({ email, role })
+        const res = await updateUserRole({ email, role })
+        console.log(res);
+        if (res?.modifiedCount === 1) {
+            toast.success('Role Updated')
+        }
     }
 
-    const handleDeleteUser = (email) => {
-        console.log(email)
-        deleteUser(email)
-    }
+
+    const handleDeleteUser = async (email) => {
+        Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showCancelButton: true,
+            confirmButtonColor: "#d33",
+            cancelButtonColor: "#3085d6",
+            confirmButtonText: "Yes, delete it!",
+        }).then(async (result) => {
+            if (result.isConfirmed) {
+                const response = await deleteUser(email);
+                if (response.data.deletedCount > 0) {
+                    Swal.fire("Deleted!", "The user has been deleted.");
+                }
+            }
+        });
+    };
+
     return (
         <tr className="w-full text-neutral-600 font-medium border-b hover:bg-slate-50">
             <td className="p-2">{index + 1}</td>
@@ -34,7 +59,15 @@ const UsersData = ({ userData, index }) => {
             </td>
 
             <td className="p-2">
-                <button onClick={() => handleDeleteUser(email)} title="Delete User"><RiDeleteBin6Fill className="text-red-600 p-1 border rounded-md border-red-600 hover:text-white hover:bg-red-600" size={25} /></button>
+                <button
+                    onClick={() => handleDeleteUser(email)}
+                    title="Delete User"
+                    disabled={role === 'admin' && user?.email === email}
+                    className={`text-red-600 p-1 border rounded-md border-red-600 hover:text-white hover:bg-red-600 ${role === 'admin' && user?.email === email ? 'opacity-50 cursor-not-allowed' : ''
+                        }`}
+                >
+                    <RiDeleteBin6Fill size={20} />
+                </button>
             </td>
         </tr>
     );
